Extract final clean data using either absolute or best method generated outliers.

Extract final clean data using either absolute or best method generated outliers.

Usage

clean_data(
  data,
  outliers,
  sp = NULL,
  mode = "best",
  colsp = NULL,
  threshold = NULL,
  verbose = FALSE,
  warn = FALSE,
  pabs = 0.1,
  autothreshold = FALSE
)

Arguments

data

The reference data for the species used in outlier detection.

outliers

Output from the outlier detection process.

sp

species name. NULL if a single species is considered in outlier detection.

mode

Either absolute to use absolute outliers to filter data or bestmethod to outliers from best method

colsp

A parameter to be used if the data is a data frame and the user must indicate the column wih species names.

threshold

value to consider whether the outlier is an absolute outlier or not.

verbose

Produces messages or not. Default FALSE. ###param multiple TRUE for multiple species and FALSE for single species considered during outlier detection.

warn

If FALSE, warning on whether absolute outliers obtained at a low threshold is indicated. Default TRUE.

pabs

Percentage of outliers allowed to be extracted from the data. If best is used to extract outliers and the pabs is exceeded, the absolute outliers are removed instead. This because some records in the best methods are repeated and they will likely to remove true values as outliers.

autothreshold

Identifies the threshold with mean number of absolute outliers.The search is limited within 0.51 to 1 since thresholds less than are deemed inappropriate for identifying absolute outliers. The autothreshold is used when threshold is set to NULL.

Value

cleaned data frame of species records.