Flags outliers or no outliers for multiple outlier methods.

Flags outliers or no outliers for multiple outlier methods.

Usage

detect(
  x,
  var,
  output,
  exclude,
  optpar,
  kmpar,
  kmedoidpar,
  ifpar,
  lofpar,
  jkpar,
  gloshpar,
  mahalpar,
  knnpar,
  zpar,
  methods,
  verbose,
  spname,
  warn,
  missingness,
  showErrors
)

Arguments

x

Dataframe with species occurrences

var

Environmental parameter considered in flagging suspicious outliers.

output

Either clean; to produce a data set with no outliers, or outlier: to output a dataframe with outliers. Default outlier.

exclude

Exclude variables that should not be considered in the fitting the one class model, for example x and y columns or latitude/longitude or any column that the user doesn't want to consider.

optpar

Parameters for species optimal ranges like temperatures ranges multidetect.

kmpar

Parameters for kmeans clustering like method and number of clusters for tuning multidetect.

kmedoidpar

Parameters for adjusting the distance metrics used in kmedoid method multidetect.

ifpar

Isolation forest parameter settings. Parameters of the isolation model that are required, include the cutoff to be used for denoting outliers. It ranges from 0 to 1 but Default 0.5. Also, the size of data partitioning for training should be determined. For more details check (Liu et al. 2008)

lofpar

Parameters for local outlier factor such as the distance matrix and mode of method implementation such as robust and soft modes multidetect.

jkpar

Parameters for reverse jack knifing mainly the mode used multidetect.

gloshpar

Parameters for global local outlier score from hierarchies such as distance metric used multidetect.

mahalpar

Parameters for Malahanobis distance which includes varying the mode of output multidetect.

knnpar

Parameters for varying the distance matrix such euclidean or Manhattan multidetect.

zpar

Parameters for z-score such mode and x parameter multidetect.

methods

Outlier detection methods considered. Use extractMethod to get outliers implemented in the package.

verbose

whether to return messages or not. Default FALSE.

spname

species name being handled.

warn

whether to return warning or not. Default TRUE.

missingness

Allowed missing values in a column to allow a user decide whether to remove the individual columns or rows from the data sets. Default 0.1. Therefore, if if a column has more than 10% missing values, then it will be removed from the dataset rather than the rows.

showErrors

show execution errors and therefore for multiple species the code will break if one of the methods fails to execute.

Value

Dataframe with or with no outliers.

References

Liu FeiT, Ting KaiM, Zhou Z-H. 2008. Isolation Forest. Pages 413-422 In 2008 Eighth IEEE International Conference on Data Mining. Available from https://ieeexplore.ieee.org/abstract/document/4781136 (accessed November 18, 2023).

See also