Determines the species ranges, whether native or alien based on Freshwater Information Platform.

Determines the species ranges, whether native or alien based on Freshwater Information Platform.

Usage

endemicity(basin, range)

Arguments

basin

Basin under consideration.

range

Whether introduced (a) or native (n).

Value

dataframe showing species ranges

Examples

if (FALSE) { # \dontrun{

df <- endemicity(basin ='Danu', range =c('a','n'))
} # }