An internal function to check for missing value in environmental data before extraction.

An internal function to check for missing value in environmental data before extraction.

Usage

exdata(raster, coords, labels, missingness, exclude, vifcutoff, verbose)

Arguments

raster

raster. A raster layer with species environmental data to be used during modelling.

coords

coordiantes. the coordinates geometry. Either x and y or geometry column if the species occurrences are in a geometry format.

labels

string. If the dataset has presence absence column, it should be indicated here to be used for extracting the labels. Character labels for example allowed, for example P for presences and A for absences recommended.

missingness

numeric. The limit at which a column is discarded from the data rather than removing numerous rows. For example, if the column has 50% of the dataset missing values then removing the rows will be problematic rather than column since it only affect one parameter than the whole dataset.

exclude

vector. Indicate columns where checking non missing values or NAs should not be done.

vifcutoff

numeric. Used in assessing multicolinearity in environmental predictors using correlation from vifcor function from usdm package Naimi et al., 2014.

verbose

logical. To return execution messages or not. The Default is FALSE.

Value

extracted environmental data with labels.

References

Naimi, B., Hamm, N. A., Groen, T. A., Skidmore, A. K., & Toxopeus, A. G. (2014). Where is positional uncertainty a problem for species distribution modelling?. Ecography, 37(2), 191-203.