Extract basin ranges from Freshwater Information Platform.

Extract basin ranges from Freshwater Information Platform.

Usage

fip_ranges(data, colsp = NULL, basin, range, pct = 90, verbose = F, sn = FALSE)

Arguments

data

Data frame to extract basin species records which are either native, endemic or both.

colsp

A variable in the data with species names

basin

Basin names as indicated in the Freshwater Information Platform (Schmidt-Kloiber et al., 2019)

range

For each basin, either indicate n to return only native, a: alien (Schmidt-Kloiber et al., 2019)

pct

The percentage similarity to be considered to assign a relatively similar name from FishBase if the exact match is absent. The higher the pct values, the higher the different similar are considered to replace a species name that is cheeked from FisBase.

verbose

TRUE to return processing messages or FALSE for no messages. Default FALSE

sn

Either FALSE to return accepted names from FishBase rather than synonyms.Default FALSE.

Value

Species data frame with temperature ranges from FishBase

References

Schmidt-Kloiber, A., Bremerich, V., De Wever, A. et al. The Freshwater Information Platform: a global online network providing data, tools and resources for science and policy support. Hydrobiologia 838, 1-11 (2019).

Examples

if (FALSE) { # \dontrun{

data(jdsdata)

nativedanube <- fip_ranges(data = jdsdata, colsp = 'speciesname',
                           range = 'n', verbose = T, basin = 'Danu')

#Species introduced in Danube

aliendf <- fip_ranges(data = jdsdata, colsp = 'speciesname', basin = 'Danu', range = 'a')

} # }