Sequential fences constants

A tibble data with k constants for sequential fences method.

Usage

data(kdat)

Format

A tibble 101 rows and 2 columns.

Details

k constants fro flagging outliers with several chnages in the fences.

References

Schwertman NC, de Silva R. 2007. Identifying outliers with sequential fences. Computational Statistics and Data Analysis 51:3800–3810.

Examples

data("kdat")
kdat
#>       n      kn
#> 1     5 1.65798
#> 2     6 1.28351
#> 3     7 1.51475
#> 4     8 1.32505
#> 5     9 1.50427
#> 6    10 1.31212
#> 7    11 1.45768
#> 8    12 1.32968
#> 9    13 1.45268
#> 10   14 1.32353
#> 11   15 1.42975
#> 12   16 1.33318
#> 13   17 1.42684
#> 14   18 1.32959
#> 15   19 1.41322
#> 16   20 1.33568
#> 17   21 1.41132
#> 18   22 1.33333
#> 19   23 1.40230
#> 20   24 1.33753
#> 21   25 1.40096
#> 22   26 1.33587
#> 23   27 1.39455
#> 24   28 1.33894
#> 25   29 1.39355
#> 26   30 1.33770
#> 27   31 1.38876
#> 28   32 1.34004
#> 29   33 1.38799
#> 30   34 1.33909
#> 31   35 1.38428
#> 32   36 1.34092
#> 33   37 1.38367
#> 34   38 1.34017
#> 35   39 1.38071
#> 36   40 1.34165
#> 37   41 1.38021
#> 38   42 1.34104
#> 39   43 1.37779
#> 40   44 1.34226
#> 41   45 1.37737
#> 42   46 1.34175
#> 43   47 1.37536
#> 44   48 1.34278
#> 45   49 1.37501
#> 46   50 1.34235
#> 47   51 1.37331
#> 48   52 1.34322
#> 49   53 1.37301
#> 50   54 1.34285
#> 51   55 1.37156
#> 52   56 1.34361
#> 53   57 1.37130
#> 54   58 1.34329
#> 55   59 1.37004
#> 56   60 1.34394
#> 57   61 1.36981
#> 58   62 1.34366
#> 59   63 1.36871
#> 60   64 1.34424
#> 61   65 1.36851
#> 62   66 1.34399
#> 63   67 1.36754
#> 64   68 1.34450
#> 65   69 1.36737
#> 66   70 1.34429
#> 67   71 1.36650
#> 68   72 1.34474
#> 69   73 1.36635
#> 70   74 1.34454
#> 71   75 1.36557
#> 72   76 1.34495
#> 73   77 1.36543
#> 74   78 1.34478
#> 75   79 1.36474
#> 76   80 1.34514
#> 77   81 1.36461
#> 78   82 1.34499
#> 79   83 1.36398
#> 80   84 1.34532
#> 81   85 1.36387
#> 82   86 1.34517
#> 83   87 1.36330
#> 84   88 1.34548
#> 85   89 1.36319
#> 86   90 1.34535
#> 87   91 1.36267
#> 88   92 1.34562
#> 89   93 1.36258
#> 90   94 1.34550
#> 91   95 1.36210
#> 92   96 1.34576
#> 93   97 1.36201
#> 94   98 1.34565
#> 95   99 1.36157
#> 96  100 1.34588
#> 97  200 1.34740
#> 98  300 1.34792
#> 99  400 1.34818
#> 100   0 1.34898