Basin ranges for multiple species.

Basin ranges for multiple species.

Usage

mbasinranges(
  occurrences,
  colsp = NULL,
  lat,
  lon,
  verbose = TRUE,
  discard = TRUE,
  batch = TRUE,
  output = "records"
)

Arguments

occurrences

dataframe. Species occurrences for a particular to check for the basin where they lie in.

colsp

character. Species column considered as its recorded in the occurrences.

lat, lon

coordinates. Latitude and longitude column names in the species occurrences records.

verbose

logical. Whether implementation messages are shown. Default TRUE.

discard

logical. Whether to discard records with basins or not. Default FALSE.

batch

logical. if TRUE all species records will be considered at once. Otherwise species-wise checks can can be done. Default TRUE

output

character. Whether to produce the records or the basin files. Default records.

Value

species occurrence records and their basins were they are found.

Examples

if (FALSE) { # \dontrun{

data("efidata")

salmo1 <- efidata[efidata[,'scientificName'] == c("Anguilla anguilla", "Cottus gobio"), ]

br <- mbasinranges(occurrences = salmo1, lat = 'decimalLatitude', lon = 'decimalLongitude')

} # }