Determine species distribution ranges based on IUCN RedList.

Determine species distribution ranges based on IUCN RedList.

Usage

mdistr_ranges(data, colsp, lat, lon, key, mode, verbose = T)

Arguments

data

dataframe. species data frame to check with International Union for Conservation of Nature.

colsp

string. The column names with species data.

lat, lon

coordinates. Columns with latitude and longitude in the species data frame.

key

string. API key to access IUCN RedList data. Use rl_use_iucn() to apply for it.

mode

string. indicating whether the distribution range dist of the species should be collated or species IUCN status iucn or both.

verbose

logical. TRUE to return messages or warning during the processing of data. Default is FALSE

Value

species records within the known distribution ranges of the species according to IUCN RedList database for multiple species.

Examples

if (FALSE) { # \dontrun{

mdst <- 12

} # }