Comparing the model performance before and after outlier removal.

The function is used to compare model performance before and outlier removal. The different thresholds are indicated and the model performance outputs are obtained based on threshold dependent and independent metrics such as accuracy, sensitivity, specificity, true skill statistics, and Area Under the Curve.

Usage

modelcomparison(
  refdata,
  outliers,
  raster,
  models = "GLM",
  metrics = "all",
  thresholds = NULL,
  loessthreshold = FALSE,
  autothreshold = FALSE,
  mode = "abs",
  colsp = NULL,
  nboots = 10,
  testprop = 0.2,
  geom = NULL,
  lat = NULL,
  lon = NULL,
  prbackground = 1,
  binary = FALSE,
  verbose = F,
  warn = F,
  full = FALSE,
  pabs = 0.1,
  minpts = 10,
  plot = FALSE,
  ...
)

Arguments

refdata

list or dataframe of species data generated during pre-cleaning process. This data set is used as the reference data in accessing the outlier detection performance using both threshold dependent and independent metrics. This data set must be same as when in detecting outliers to avoid mismatch.

outliers

datacleaner class data set with all outliers obtained using the multdetect function.

raster

SpatRaster. The environmental where pseudo absences are extracted using envextract function.

models

character. The models to be used to examine the relationship between species occurrences and environmental parameters. Only Random forest and generalized linear models are accepted. GLM is used to set to Generalized Linear Models and RF for Random Forest. RF1 variant is slower and is suggested if RF fails.

metrics

character. Either to only consider threshold-dependent dep, threshold-independent indep or all evaluation metrics all to determine model performance.

thresholds

numeric. The thresholds ranging from 0.1 to 1 but greater than 0.5 is advisable.

loessthreshold

whether to consider loess or not.

autothreshold

logical. Identifies the threshold with mean number of absolute outliers.The search is limited within 0.51 to 1 since thresholds less than are deemed inappropriate for identifying absolute outliers. The autothreshold is used when threshold is set to NULL.

mode

character of either abs to use absolute outliers to filter data or best to use the best method to filter the data bestmethod function.

colsp

character. Column with species names if the species occurrence is a data frame not list,

nboots

numeric Creating sub samples for modeling evaluation. At least more than 5 is advisable.

testprop

numeric. The proportion to be used for partitioning data. Ranges from 0.1 to 1. Less than 0.5 is advisable to allow have more model training data.

geom

character. Column name used to indicate geometry column during data extraction when the species occurrences is of sf class and have geometry column instead of latitude and longitude.

lat, lon

character. Column names for latitude/longitude if the species occurrences don't have the geometry column or not spatial vector file. The latitude and longitude must be provided for data extraction from the raster layers.

prbackground

numeric. Value indicating the proportion of pseudo absences extracted in relation to the species presences. Used in spatSample function. Ranges from at least 0.1 to 1. Check (Barbet-Massin et al., 2012) on the number of pseudo absences required during species distribution modeling.

binary

logical. If FALSE the species dataset do not have label column for presence absence and label is NULL. If only species presences and absences are indicated, the user should select TRUE.

verbose

logical. Return messages or not. Default is FALSE.

warn

logical. Return warning or not. Default is FALSE.

full

logical. Either full model output for TRUE or FALSE for only performance metrics output.

pabs

numeric. Percentage of outliers allowed to be extracted from the data. If best is used to extract outliers and the pabs is exceeded, the absolute outliers are removed instead. This because some records in the best methods are repeated and they will likely to remove true values as outliers.

minpts

numeric value indicating the minimum number of records required to run the species distribution model. Default is 10.

plot

true or false

...

Other parameters can be used from envextract function.

Value

A list of model instance and model performance metrics.

References

Barbet-Massin, M., Jiguet, F., Albert, C. H., & Thuiller, W. (2012). Selecting pseudo-absences for species distribution models: how, where and how many? Methods Ecol Evol 3: 327–338.

See also

Examples

if (FALSE) {

#worldclim data for to extract environmental predictors
worldclim <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

#get 500 records online using getdata function to compliment salmo trutta records and basin polygon
#basin to get the bounding box to delineate the area of concern

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'), quiet=TRUE)

salmonline <- getdata(data = "Salmo trutta", gbiflim = 500, inatlim = 3, vertlim = 3, bbox = db)

salextract <- extract_online(salmonline)

#merge both online and offline data and filter Salmo trutta

#select species with enough records

datafinal <- salextract[salextract[,'species'] == "Salmo trutta", ]

#initial data extraction and preliminary analysis

rdata <- pred_extract(data = datafinal,
                      raster= worldclim ,
                      lat = 'decimalLatitude',
                      lon= 'decimalLongitude',
                      colsp = 'species',
                      bbox  = db,
                      multiple = FALSE,
                      minpts = 10,
                      list=TRUE,
                      merge=F, verbose = F)

#apply ensemble outlier detection. Note: (x and y in exclude parameter are
#internally generally in pred_extract during environmental data extraction )

outliersdf <- multidetect(data = rdata, multiple = FALSE,
                          var = 'bio6',
                          output = 'outlier',
                          exclude = c('x','y'),
                          methods = c('zscore', 'adjbox','iqr', 'semiqr','hampel'))

modeout <- modelcomparison(refdata = rdata, outliers = outliersdf, raster = worldclim,
                           lat = 'y', lon = 'x', models = c("GLM"),
                           mode = 'best', testprop = 0.2, metrics = 'all',
                           thresholds = 0.2, full = FALSE, minpts = 10)

getper <- get_performance(modeloutput = modeout)

ggperform(modelout = modeout, cutoff = 0.1)

}