Identifies best outlier detection method suing Sorensen Similarity Index.

Identifies best outlier detection method suing Sorensen Similarity Index.

Usage

sorensen(x, sp = NULL, threshold = NULL, warn = FALSE, autothreshold = FALSE)

Arguments

x

List of data frames for each methods used to identify outliers in mult_detect function.

sp

species name or index if multiple species are considered during outlier detection.

threshold

Maximum value to denote an absolute outlier. The threshold ranges from 0 which indicates a point has not been flagged by any outlier detection method as an outlier or 1, when means the record is an absolute or true outlier sicen it has been identified by all methods. At both extremes, at low threshold values, many records are classified, which may be due to individual method weakness or strength and data distribution. Also, at higher threshold values, the true outliers are retained Fo example, if 10 methods are considered and 9 methods flags a record as an outlier, If a cut off 1 is used, then that particular record is retained. Therefore the default cutoff is 0.6 but autothreshold can be used to select the appropriate threshold.

warn

If TRUE, warning on whether absolute outliers obtained at a low threshold is indicated. Default TRUE.

autothreshold

Identifies the threshold with mean number of absolute outliers.The search is limited within 0.51 to 1 since thresholds less than are deemed inappropriate for identifying absolute outliers. The autothreshold is used when threshold is set to NULL.

Value

best method for identifying outliers.

Examples

if (FALSE) {

data(efidata)

db <- sf::read_sf(system.file('extdata/danube/basinfinal.shp', package = "specleanr"), quiet = TRUE)

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package = "specleanr"))

checkname <- check_names(data=efidata, colsp ='scientificName', pct = 90, merge = T)

extdf <- pred_extract(data = checkname, raster = wcd,
                      lat = 'decimalLatitude', lon = 'decimalLongitude',
                     colsp = 'speciescheck',
                     list = TRUE,verbose = F,
                     minpts = 6,merge = F)#basin removed

 #outlier detection

outliers <- multidetect(data = extdf, var = 'bio6', output='outlier',
                       exclude = c('x','y'), multiple = TRUE,
                       methods = c('mixediqr', "iqr", "kmeans", "mahal"),
                       kmpar =list(k=6, method='silhouette', mode='soft'))

sordata <- sorensen(x = outliers, sp= 1, threshold = 0.2, var = 'bio6')#

}