Threshold dependent parameters for species distribution model evaluations.

Threshold dependent parameters for species distribution model evaluations.

Usage

dep(observed, predicted, P, A)

Arguments

observed

The observed occurrences of the species.

predicted

The fitted model obtained from sdmfit

P

Character for presence label. Important to indicate the presence/absence character to correctly compute the evaluation metric in evaluate function.

A

Character for absence label. Important to indicate the presence/absence character to correctly compute the evaluation metric in evaluate function.

Value

A list of threshold dependent metrics after models evaluation.

Details

These are threshold-dependent metrics which vary if the cutoff is changed. The default is 0.5 which indicates that probabilities greater than 0.5 are recorded as the species is present and less than 0.5 indicates absence of a species. However, if the cutoff is increased, probabilities converted as positive reduce and also the model evaluation metrics such as sensitivity, specificity or accuracy changes significantly. Different literature is used to compute the metrics including (Allouche et al. 2006) and (Erickson & Kitamura 2021).

References

Allouche O, Tsoar A, Kadmon R. 2006. Assessing the accuracy of species distribution models: Prevalence, kappa and the true skill statistic (TSS). Journal of Applied Ecology 43:1221-1232.

Erickson BJ, Kitamura F. 2021. Magician's corner: 9. performance metrics for machine learning models. Radiology: Artificial Intelligence 3:1-7.

See also