Download species records from online database.

Download species records from online database.

Usage

getdata(
  data,
  colsp = NULL,
  bbox = NULL,
  isFish = TRUE,
  gbiflim = 1e+06,
  vertlim = 1000,
  inatlim = 3000,
  verbose = FALSE,
  warn = FALSE,
  ...
)

Arguments

data

Data frame, list or vector of species names to retrieve records from online databases (GBIF, VertNET, and iNaturalist).

colsp

A variable of species names. Provided if data is a data frame, so not required for lists and vector.

bbox

Bounding box to limit the download of records within a particular area. Otherwise all records from the GBIF will be downloaded. These can be provided in two forms, either a shapefile (sf) class accepted or provide a list of xmin, ymin, xmax, and ymax.

isFish

logical, to indicate if the occurrence records extracted are for fish taxa or not. This allows to clean the species names accordingly. For other taxa a different name checks is conducted. Default is TRUE.

gbiflim

Limits on the records from the Global Biodiversity Information Platform

vertlim

Limits on the records from VertNET.

inatlim

Limits on the records from iNaturalist database.

verbose

TRUE if detailed messages should be indicated and FALSE if download messages are not needed. Default TRUE.

warn

logical, to indicate if warning messages should be shown. Default FALSE.

...

More function for species data download can be used. See rgbif::occ_data for more information, rinat::get_inat_obs, and rvertnet::searchbyterm.

Value

Lists of species records from online databases

Details

Note always check the validity of the species name with standard database FishBase or World Register of Marine Species. If the records are more than 50000 in GBIF, and extent can be provide to limit the download.

Examples

if (FALSE) {

gbaloni <- check_names(data = 'gymnocephalus baloni', pct=90)

gbdata <- getdata(data=gbaloni, gbiflim = 100, inatlim = 100, vertlim = 100)

#Get for two species
sp_records <- getdata(data=c('Gymnocephalus baloni', 'Hucho hucho'),gbiflim = 100,
                            inatlim = 100,
                            vertlim = 100)

#3. Generate species list from a dataframe

data(efidata)

danube <- system.file('extdata/danube/basinfinal.shp', package='specleanr')

danubebasin <- sf::st_read(danube, quiet=TRUE)

datachecked <- check_names(data= efidata,colsp = 'species', pct=90, merge = TRUE)#clean names

Onlinedf <- getdata(data = datachecked, colsp='speciescheck', bbox=danubebasin)

}