Allows ensemble multiple outlier detection methods.

Allows ensemble multiple outlier detection methods.

Usage

multidetect(
  data,
  var,
  output = "outlier",
  exclude = NULL,
  multiple,
  colsp = NULL,
  optpar = list(optdf = NULL, ecoparam = NULL, optspcol = NULL, direction = NULL, maxcol
    = NULL, mincol = NULL, maxval = NULL, minval = NULL, checkfishbase = FALSE, mode =
    "geo", lat = NULL, lon = NULL, pct = 80, warn = FALSE),
  kmpar = list(k = 6, method = "silhouette", mode = "soft"),
  kmedoidpar = list(k = 2, metric = "manhattan", full = FALSE),
  ifpar = list(cutoff = 0.5, size = 0.7),
  mahalpar = list(mode = "soft"),
  jkpar = list(mode = "soft"),
  zpar = list(type = "mild", mode = "soft"),
  gloshpar = list(k = 3, metric = "manhattan", mode = "soft"),
  knnpar = list(metric = "manhattan", mode = "soft"),
  lofpar = list(metric = "manhattan", mode = "soft", minPts = 10),
  methods,
  verbose = FALSE,
  spname = NULL,
  warn = TRUE,
  missingness = 0.1,
  showErrors = TRUE
)

Arguments

data

Dataframe or list of data sets for multiple or single species after after of extraction of environment predictors.

var

A variable to check for outliers especially the one with directly affects species distribution such as maximum temperature of the coldest month for bioclimatic variables (IUCN Standards and Petitions Committee, 2022)) or stream power index for hydromorphological parameters (Logez et al., 2012).

output

Either clean: for a data set with no outliers, or outlier: to output a dataframe with outliers. Default outlier.

exclude

Exclude variables that should not be considered in the fitting the one class model, for example x and y columns or latitude/longitude or any column that the user doesn't want to consider.

multiple

If the multiple species considered, then multiple set to TRUE: If FALSE then multiple should be set to FALSE.

colsp

A column with species columns if the data set for species is a dataframe not a list. see pred_extract for extracting environmental data.

optpar

Parameters for species optimal ranges like temperatures ranges.

kmpar

Parameters for k-means clustering like method and number of clusters for tuning.

kmedoidpar

Parameters for adjusting the distance metrics used in kmedoid method.

ifpar

Isolation forest parameter settings. Parameters of the isolation model that are required include the cutoff to be used for denoting outliers. It ranges from 0 to 1 but Default 0.5. Also, the size of data partitioning for training should be determined. For more details check (Liu et al. 2008)

mahalpar

Parameters for Malahanobis distance which includes varying the mode of output.

jkpar

Parameters for reverse jack knifing mainly the mode used.

zpar

Parameters for z-score such mode and x parameter.

gloshpar

Parameters for global local outlier score from hierarchies such as distance metric used..

knnpar

Parameters for varying the distance matrix such as Euclidean or Manhattan distance.

lofpar

Parameters for local outlier factor such as the distance matrix and mode of method implementation such as robust and soft modes..

methods

Outlier detection methods considered. Use extractMethod to get outliers implemented in the package.

verbose

whether to return messages or not. Default FALSE.

spname

species name being handled.

warn

whether to return warning or not. Default TRUE.

missingness

Allowed missing values in a column to allow a user decide whether to remove the individual columns or rows from the data sets. Default 0.1. Therefore, if if a column has more than 10% missing values, then it will be removed from the dataset rather than the rows.

showErrors

show execution errors and therefore for multiple species the code will break if one of the methods fails to execute.

Value

Outliers or clean dataset.

Details

This function computes different outlier detection methods including univariate, multivariate and species ecological ranges to enables seammless comaprision and similarities in the outliers detected by each method. This can be done for multiple species or a single species in a dataframe or lists or dataframes and thereafter the outliers can be extracted using the clean_data_extract function.

References

IUCN Standards and Petitions Committee. (2022). THE IUCN RED LIST OF THREATENED SPECIESTM Guidelines for Using the IUCN Red List Categories and Criteria Prepared by the Standards and Petitions Committee of the IUCN Species Survival Commission. https://www.iucnredlist.org/documents/RedListGuidelines.pdf.

Liu FeiT, Ting KaiM, Zhou Z-H. 2008. Isolation Forest. Pages 413–422 In 2008 Eighth IEEE International Conference on Data Mining. Available from https://ieeexplore.ieee.org/abstract/document/4781136 (accessed November 18, 2023).

Examples