Preliminary data cleaning including removing duplicates, records outside a particular basin, and NAs.

Preliminary data cleaning including removing duplicates, records outside a particular basin, and NAs.

Usage

pred_extract(
  data,
  raster,
  lat,
  lon,
  bbox = NULL,
  colsp,
  minpts = 10,
  multiple = TRUE,
  list = TRUE,
  merge = FALSE,
  verbose = FALSE,
  warn = FALSE
)

Arguments

data

Data frame with multiple species or only one species for checking records with no coordinates, duplicates, and check for records that fall on land, sea, country or city centroids, and geographical outliers(Zzika et al., 2022).

raster

Environmental layers from different providers such as WORLDCLIM (), Hydrogaphy90m (), CHELSA, Copernicus ().

lat

variable for latitude column name.

lon

variable for longitude column name.

bbox

Object of class 'shapefile' If only a particular basin is considered. Bounding box vector points can also be provided in the form "c(xmin, ymin, xmax, ymax)". xmin is the minimum longitude, ymin is the minimum latitude, xmax is the maximum longitude and xmax is the minimum latitude.

colsp

variable already in the data.

minpts

minimum number of records for the species after removing duplicates and those within a particular basin.

multiple

TRUE if species are more than one and FALSE for one species in the data.

list

If TRUE the a list of multiple species data frames will be generated and FALSE for a dataframe of species data sets. Default TRUE

merge

To add the other columns in the species data after data extraction. Default TRUE

verbose

if TRUE message and warnings will be produced. Default TRUE

warn

logical, indicating to whether to show implementation warning or not. Default FALSE.

Value

precleaned cleaned data sets for single or multiple species

Examples

if (FALSE) {

data(jdsdata)

danube <- system.file('extdata/danube/basinfinal.shp', package='specleanr')

danubebasin <- sf::st_read(danube, quiet=TRUE)

#Get environmental data

worldclim <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

referencedata <- pred_extract(data = jdsdata,
                          raster= worldclim ,
                          lat ='decimalLatitude',
                          lon = 'decimalLongitude',
                          colsp = 'speciesname',
                          bbox = danubebasin,
                          multiple = TRUE,
                          list= TRUE, #list will be generated for all species
                          minpts = 10, merge=T)
}